pyproject.toml
[tool.poetry]
name = "chuviscobot"
version = "0.1.0"
description = "Bot de Telegram para os grupos oficiais do Garoa Hacker Clube"
authors = ['Felipe "Juca" Sanches', "Luciana Marques", "Priscilla Gutierres"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.3"
sqlalchemy = "^2.0.16"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
ruff = "^0.0.272"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-vvv --cov-report=term-missing --cov=chuvisco"
testpaths = ["tests/unit"]
log_level = "DEBUG"
filterwarnings = []
